+++ src/esm/components/icons/align-center-icon/index.js
@@ -12,7 +12,7 @@ export var AlignCenterIcon = function AlignCenterIcon(_ref) {
     height: "10",
     viewBox: "0 0 14 10",
     xmlns: "http://www.w3.org/2000/svg"
-  }, props), title && /*#__PURE__*/React.createElement("title", null, title), ">", /*#__PURE__*/React.createElement("path", {
+  }, props), title && /*#__PURE__*/React.createElement("title", null, title), /*#__PURE__*/React.createElement("path", {
     d: "M12 8v2H3V8h9zm2-4v2H0V4h14zm-1-4v2H2V0h11z",
     fill: color
   }));